+++ always-provision.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+echo "----------------------------------------"
+echo "=== Starting the always provisioning ==="
+echo "----------------------------------------"
+
+echo "Start docker"
+sudo systemctl start docker