+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Added
+- Script always-provision.sh for always provisioning of Vagrant.
 
 ## [0.3.1] - 2015-03-01
 ### Fixed